Add Tableau provider separate from Salesforce Provider (#14030)
Closes #13614
