+++ airflow/providers/dependencies.json
@@ -70,6 +70,9 @@
   "postgres": [
     "amazon"
   ],
+  "salesforce": [
+    "tableau"
+  ],
   "sftp": [
     "ssh"
   ],
