+++ docs/spelling_wordlist.txt
@@ -1283,6 +1283,7 @@ sync'ed
 sys
 syspath
 systemd
+tableau
 tableauserverclient
 tablefmt
 tableId
